Add setting of error callback to GLFW initialization
